# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Bot-Mox API v1
  version: 1.0.0
  description: Canonical API surface for phased backend/frontend alignment.
servers:
  - url: http://localhost:3002
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/v1/ipqs/check:
    post:
      summary: Check single IP via IPQS
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ip:
                  type: string
              required: [ip]
      responses:
        '200':
          description: IPQS check result
        '400':
          description: Invalid request
  /api/v1/ipqs/check-batch:
    post:
      summary: Check batch of IPs via IPQS
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ips:
                  type: array
                  items:
                    type: string
              required: [ips]
      responses:
        '200':
          description: Batch check results
        '400':
          description: Invalid request
...
  /api/v1/infra/proxmox/status:
    get:
      summary: Infra Proxmox status (infra role)
      responses:
        '200':
          description: Infra health
        '403':
          description: Forbidden
...
  /api/v1/infra/proxmox/nodes/{node}/qemu:
    get:
      summary: List node VMs (infra role)
      parameters:
        - in: path
          name: node
          required: true
          schema:
            type: string
      responses:
        '200':
          description: VM list
  /api/v1/infra/proxmox/nodes/{node}/qemu/{vmid}/clone:
    post:
      summary: Clone VM on node (infra role)
      parameters:
        - in: path
          name: node
          required: true
          schema:
            type: string
        - in: path
          name: vmid
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                newid:
                  type: string
                name:
                  type: string
                target:
                  type: string
                full:
                  type: boolean
              additionalProperties: true
      responses:
        '200':
          description: Clone task accepted
        '400':
          description: Invalid request
  /api/v1/infra/proxmox/nodes/{node}/qemu/{vmid}/config:
    get:
      summary: Read VM config from Proxmox (infra role)
      parameters:
        - in: path
          name: node
          required: true
          schema:
            type: string
        - in: path
          name: vmid
          required: true
          schema:
            type: string
      responses:
        '200':
          description: VM config payload
    put:
      summary: Update VM config in Proxmox (infra role)
      parameters:
        - in: path
          name: node
          required: true
          schema:
            type: string
        - in: path
          name: vmid
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                cores:
                  type: integer
                memory:
                  type: integer
                sockets:
                  type: integer
              additionalProperties: true
      responses:
        '200':
          description: VM config updated
        '400':
          description: Invalid request
  /api/v1/infra/proxmox/nodes/{node}/tasks/{upid}/status:
    get:
      summary: Get Proxmox task status (infra role)
      parameters:
        - in: path
          name: node
          required: true
          schema:
            type: string
        - in: path
          name: upid
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task status payload
  /api/v1/infra/proxmox/nodes/{node}/qemu/{vmid}/status/{action}:
    post:
      summary: Execute VM action (start/stop/reboot/...) (infra role)
      parameters:
        - in: path
          name: node
          required: true
          schema:
            type: string
        - in: path
          name: vmid
          required: true
          schema:
            type: string
        - in: path
          name: action
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Action result
  /api/v1/infra/proxmox/nodes/{node}/qemu/{vmid}:
    delete:
      summary: Delete VM on node (infra role)
      parameters:
        - in: path
          name: node
          required: true
          schema:
            type: string
        - in: path
          name: vmid
          required: true
          schema:
            type: string
        - in: query
          name: purge
          required: false
          schema:
            type: string
        - in: query
          name: destroy-unreferenced-disks
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Delete result
  /api/v1/infra/proxmox/nodes/{node}/qemu/{vmid}/sendkey:
    post:
      summary: Send keyboard key to VM (infra role)
      parameters:
        - in: path
          name: node
          required: true
          schema:
            type: string
        - in: path
          name: vmid
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                key:
                  type: string
              required: [key]
      responses:
        '200':
          description: Sendkey result
        '400':
          description: Invalid request
  /api/v1/infra/proxmox/nodes/{node}/qemu/{vmid}/status/current:
    get:
      summary: Get current VM runtime status (infra role)
      parameters:
        - in: path
          name: node
          required: true
          schema:
            type: string
        - in: path
          name: vmid
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Current VM status
...
  /api/v1/infra/ssh/exec:
    post:
      summary: Execute allowlisted SSH command (infra role)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                command:
                  type: string
                timeout:
                  type: number
      responses:
        '200':
          description: SSH command result
        '403':
          description: Forbidden or command not allowlisted
...
  /api/v1/infra/ssh/vm-config/{vmid}:
    get:
      summary: Read VM config over SSH (infra role)
      parameters:
        - in: path
          name: vmid
          required: true
          schema:
            type: string
      responses:
        '200':
          description: VM config text
    put:
      summary: Write VM config over SSH (infra role)
      parameters:
        - in: path
          name: vmid
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
              required: [content]
      responses:
        '200':
          description: VM config write result
  # --- Agents ---
...